fixed window warnings on metroidvania project

+++ projects/metroidvania/metroidvania.c
@@ -10,7 +10,7 @@
 
 internal_function void winsizecbk(int width, int height)
 {
-    create_orthographic_projection(&main_cam.proj_matrix, 0, width, 0, height, -0.1f, 100.f);
+    create_orthographic_projection(&main_cam.proj_matrix, 0.f, (float)width, 0.f, (float)height, -0.1f, 100.f);
     set_projection_matrix(&main_cam.proj_matrix.m11);
 }
 
@@ -33,7 +33,7 @@ void app_specific_init(void)
 void app_specific_update(double dt)
 {
     player_update(dt, &player, &controller);
-    camera_controls(dt, &main_cam);
+    camera_controls((float)dt, &main_cam);
 
     const v2i mouse_posi = get_mouse_position_from_system();
     mouse_follow.pos = (v2f){(float)mouse_posi.x, (float)mouse_posi.y};
